server/requirements.txt
Django==1.11.15
django-bootstrap4==0.0.7
django-cors-headers==2.4.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
gunicorn==19.6.0
pandas==0.23.4
python-docx==0.8.7
Pillow==5.3.0
opencv-python==3.4.2.17
pytesseract==0.2.4
protobuf==3.6.0
tensorflow==1.10.0
tensorflow-serving-api==1.10.0
tensorflow-serving-api-python3==1.8.0
img2pdf==0.3.3
pdf2image==1.2.1
numpy==1.16.2
